Strict null checks @/contentScript/pageEditor/dynamic (#8392)
* separate ensureContextMenu

* fix imports

* moves rest of exported backgroudn contextMenu functions to separate files

* fix tests

* overlay

* move remaining files

* fix imports

* separate ensureContextMenu

* fix imports

* moves rest of exported backgroudn contextMenu functions to separate files

* fix tests

* overlay

* move remaining files

* fix imports
