+++ src/contentScript/messenger/api.ts
@@ -48,8 +48,7 @@ export const runRendererBlock = getMethod("RUN_RENDERER_BLOCK");
 export const clearDynamicElements = getNotifier("CLEAR_DYNAMIC_ELEMENTS");
 export const updateDynamicElement = getNotifier("UPDATE_DYNAMIC_ELEMENT");
 export const runExtensionPointReader = getMethod("RUN_EXTENSION_POINT_READER");
-export const enableOverlay = getMethod("ENABLE_OVERLAY");
-export const disableOverlay = getMethod("DISABLE_OVERLAY");
+
 export const getInstalledExtensionPoints = getMethod(
   "INSTALLED_EXTENSION_POINTS",
 );
