+++ src/contentScript/messenger/registration.ts
@@ -32,13 +32,6 @@ import {
   removePersistedExtension,
 } from "@/contentScript/lifecycle"; // 202 strictNullCheck errors
 import {
-  clearDynamicElements,
-  disableOverlay,
-  enableOverlay,
-  runExtensionPointReader,
-  updateDynamicElement,
-} from "@/contentScript/pageEditor/dynamic"; // 205 strictNullCheck errors
-import {
   runBlockPreview,
   resetTab,
   runRendererBlock,
@@ -51,6 +44,9 @@ import {
 } from "@/contentScript/pipelineProtocol"; // Depends on background/messenger to pass strictNullCheck
 import { getCopilotHostData } from "@/contrib/automationanywhere/SetCopilotDataEffect"; // Depends on background/messenger to pass strictNullCheck
 import { showBannerFromConfig } from "@/contentScript/integrations/deferredLoginController"; // Depends on background/messenger to pass strictNullCheck
+import { clearDynamicElements } from "@/contentScript/pageEditor/dynamic/clearDynamicElements"; // 201 strictNullCheck errors
+import { runStarterBrickReader } from "@/contentScript/pageEditor/dynamic/runStarterBrickReader"; // 193 strictNullCheck errors
+import { updateDynamicElement } from "@/contentScript/pageEditor/dynamic/updateDynamicElement"; // 199 strictNullCheck errors
 
 expectContext("contentScript");
 
@@ -66,9 +62,8 @@ declare global {
 
     CLEAR_DYNAMIC_ELEMENTS: typeof clearDynamicElements;
     UPDATE_DYNAMIC_ELEMENT: typeof updateDynamicElement;
-    RUN_EXTENSION_POINT_READER: typeof runExtensionPointReader;
-    ENABLE_OVERLAY: typeof enableOverlay;
-    DISABLE_OVERLAY: typeof disableOverlay;
+    RUN_EXTENSION_POINT_READER: typeof runStarterBrickReader;
+
     INSTALLED_EXTENSION_POINTS: typeof getActiveExtensionPoints;
     ENSURE_EXTENSION_POINTS_INSTALLED: typeof ensureInstalled;
 
@@ -96,9 +91,8 @@ export default function registerMessenger(): void {
 
     CLEAR_DYNAMIC_ELEMENTS: clearDynamicElements,
     UPDATE_DYNAMIC_ELEMENT: updateDynamicElement,
-    RUN_EXTENSION_POINT_READER: runExtensionPointReader,
-    ENABLE_OVERLAY: enableOverlay,
-    DISABLE_OVERLAY: disableOverlay,
+    RUN_EXTENSION_POINT_READER: runStarterBrickReader,
+
     INSTALLED_EXTENSION_POINTS: getActiveExtensionPoints,
     ENSURE_EXTENSION_POINTS_INSTALLED: ensureInstalled,
