+++ src/contentScript/messenger/strict/registration.ts
@@ -51,6 +51,10 @@ import { getAttributeExamples } from "@/contentScript/pageEditor/elementInformat
 import selectElement from "@/contentScript/pageEditor/selectElement";
 import { insertPanel } from "@/contentScript/pageEditor/insertPanel";
 import { insertButton } from "@/contentScript/pageEditor/insertButton";
+import {
+  disableOverlay,
+  enableOverlay,
+} from "@/contentScript/pageEditor/dynamic/overlay";
 
 declare global {
   interface MessengerMethods {
@@ -85,6 +89,8 @@ declare global {
     SELECT_ELEMENT: typeof selectElement;
     INSERT_PANEL: typeof insertPanel;
     INSERT_BUTTON: typeof insertButton;
+    ENABLE_OVERLAY: typeof enableOverlay;
+    DISABLE_OVERLAY: typeof disableOverlay;
   }
 }
 export default function registerMessenger(): void {
@@ -120,5 +126,7 @@ export default function registerMessenger(): void {
     SELECT_ELEMENT: selectElement,
     INSERT_PANEL: insertPanel,
     INSERT_BUTTON: insertButton,
+    ENABLE_OVERLAY: enableOverlay,
+    DISABLE_OVERLAY: disableOverlay,
   });
 }
