+++ src/pageEditor/fields/SelectorSelectorWidget.tsx
@@ -29,7 +29,12 @@ import CreatableAutosuggest, {
 import SelectorListItem from "@/pageEditor/fields/selectorListItem/SelectorListItem";
 import { type Framework } from "@/pageScript/messenger/constants";
 import { useField } from "formik";
-import { disableOverlay, enableOverlay } from "@/contentScript/messenger/api";
+import {
+  disableOverlay,
+  enableOverlay,
+  cancelSelect,
+  selectElement,
+} from "@/contentScript/messenger/strict/api";
 import { type SelectMode } from "@/contentScript/pageEditor/types";
 import { useSelector } from "react-redux";
 import { type SettingsState } from "@/store/settings/settingsTypes";
@@ -44,10 +49,6 @@ import {
 import WorkshopMessageWidget from "@/components/fields/schemaFields/widgets/WorkshopMessageWidget";
 import { type ElementInfo } from "@/utils/inference/selectorTypes";
 import { inspectedTab } from "@/pageEditor/context/connection";
-import {
-  cancelSelect,
-  selectElement,
-} from "@/contentScript/messenger/strict/api";
 
 interface ElementSuggestion extends SuggestionTypeBase {
   value: string;
