+++ src/pageEditor/sidebar/DynamicModComponentListItem.tsx
@@ -27,8 +27,11 @@ import {
   UnsavedChangesIcon,
 } from "@/pageEditor/sidebar/ExtensionIcons";
 import { type UUID } from "@/types/stringTypes";
-import { disableOverlay, enableOverlay } from "@/contentScript/messenger/api";
-import { updateSidebar } from "@/contentScript/messenger/strict/api";
+import {
+  disableOverlay,
+  enableOverlay,
+  updateSidebar,
+} from "@/contentScript/messenger/strict/api";
 import { openSidePanel } from "@/utils/sidePanelUtils";
 import cx from "classnames";
 import reportEvent from "@/telemetry/reportEvent";
